Fix typings
Former-commit-id: 7a70437f0c098ff2636de5fdfbdf73b89c5e337a

+++ src/Engines/engine.ts
@@ -3708,7 +3708,7 @@ declare type RenderTargetTexture = import("Materials/Textures/renderTargetTextur
                 return;
             }
 
-            this._gl.uniformMatrix4fv(uniform, false, matrix.toArray());
+            this._gl.uniformMatrix4fv(uniform, false, matrix.toArray() as Float32Array);
         }
 
         /**
